requirements.txt
gspread==5.4.0
oauth2client==4.1.3
python-dotenv==0.20.0
pygame==2.1.2
selenium==4.4.3